Create Exercise 3: Calculator
Write a program that will display the following calculator menu

1. Add
2. Subtract
3. Multiply
4. Divide
5. Modulus
6. Check greater number
The program will prompt the user to choose the operation choice (from 1 to 6). Then it asks the user to input values for the calculation. The program outputs the results of the calculation.Use operator module functions

Extension Problem (Bonus):
Allow the user to keep entering values until they enter Q to quit.
Handle incorrect input